Removed acts_as_taggable_on from the Gemfile and the Assignee model
Purpose of this branch: To replace the simple tagging engine in the
current branch with a more comprehensive and logical solution. By
making a Task model that is related to the Assignee model in a
one-to-many fashion, I can provide understandable data while
simultaneously improving searches.

While the tagging model is a nice and simple way to add/edit Tasks (I
actually might keep the text-area field currently used to enter tasks,
since it would save the user a lot of time when entering data the first
time. I'd probably have to tweak it for editing, but I could address
that issue at a later time).

This commit removed the acts_as_taggable_on gem from the Gemfile, and
removed the definitions from the Assignee model.