UI Tweaks
Drastically improved the UI of the application, and began the first
steps in getting Kaminari setup as the pagination engine of the system.
However, I ran into a bug (reported here:
https://github.com/amatsuda/kaminari/issues/145#issuecomment-2820889 )
and will be monitoring its progress.

The last few things I want to implement in terms of UI: AJAX Searching
(Possible Feature?), Improved Header (Moving Menu over to the right),
Validation in Entry/Editing, Linking Directly to an Assignee